FastSpeech2/requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
bidict==0.23.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==14.0
contourpy==1.2.1
cycler==0.12.1
Distance==0.1.3
dnspython==2.3.0
editdistance==0.8.1
et-xmlfile==1.1.0
eventlet==0.36.1
filelock==3.14.0
Flask==2.2.5
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
flask-talisman==1.1.0
fonttools==4.51.0
fsspec==2024.5.0
g2p==2.0.0
g2p-en==2.1.0
greenlet==3.0.3
h11==0.14.0
humanfriendly==10.0
idna==3.7
inflect==7.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
more-itertools==10.2.0
mpmath==1.3.0
munkres==1.1.4
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
panphon==0.20.0
pillow==10.3.0
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
pypinyin==0.51.0
python-dateutil==2.9.0.post0
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
scipy==1.13.0
simple-websocket==1.0.0
six==1.16.0
sympy==1.12
text-unidecode==1.3
torch==2.3.0
tqdm==4.66.4
typeguard==4.2.1
typing_extensions==4.11.0
unicodecsv==0.14.1
Unidecode==1.3.8
urllib3==2.2.1
Werkzeug==3.0.3
wsproto==1.2.0
